+++ libft/srcs/num/ft_sum.c
@@ -6,23 +6,23 @@
 /*   By: zwang <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/28 15:09:30 by zwang             #+#    #+#             */
-/*   Updated: 2018/09/28 16:24:15 by zwang            ###   ########.fr       */
+/*   Updated: 2019/02/24 21:11:24 by Zexi Wang        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_sum(int total_num, ...)
+long	ft_sum(int total_num, ...)
 {
 	va_list	num_list;
 	int		i;
-	int		sum;
+	long	sum;
 
 	va_start(num_list, total_num);
 	sum = 0;
 	i = 0;
 	while (i++ < total_num)
-		sum += va_arg(num_list, int);
+		sum += va_arg(num_list, long);
 	va_end(num_list);
 	return (sum);
 }
